+++ src/app.ts
@@ -1,14 +1,12 @@
 import express from 'express';
 import swaggerUI from 'swagger-ui-express';
-import { join, dirname } from 'path';
-import { fileURLToPath } from 'url';
+import { join } from 'path';
 import YAML from 'yamljs';
 import userRouter from './resources/users/user.router.js';
 import boardRouter from './resources/boards/board.router.js';
 import taskRouter from './resources/tasks/task.router.js';
 import { boardService } from './resources/boards/board.service.js';
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
 const app = express();
 const swaggerDocument = YAML.load(join(__dirname, '../doc/api.yaml'));
 
@@ -30,6 +28,7 @@ app.param('boardId', async (req, res, next, boardId) => {
     if(!board) {
         res.sendStatus(404);
     } else {
+        //@ts-ignore
         req.board = board;
 
         if(req.body) {
